+++ ppdc/ppdc.h
@@ -1,7 +1,7 @@
 //
 // Definitions for the CUPS PPD Compiler.
 //
-// Copyright 2007-2009 by Apple Inc.
+// Copyright 2007-2019 by Apple Inc.
 // Copyright 2002-2007 by Easy Software Products.
 //
 // These coded instructions, statements, and computer programs are the
@@ -109,7 +109,7 @@ class ppdcArray				//// Shared Array
 {
   public:
 
-  int		count,			// Number of elements
+  size_t	count,			// Number of elements
 		alloc,			// Allocated elements
 		current;		// Current element
   ppdcShared	**data;			// Elements
